requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-crispy-forms==1.14.0
django-mysql==4.7.1
django-rest-framework==0.1.0
django-restframework==0.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
